chore - codegen: extract c string escaping to a separate file
